moved the code to guide_sidebar
